gateway/requirement.txt
fastapi
strawberry-graphql
uvicorn
python-dotenv
grpcio
protobuf
grpcio-tools